// view/next-project/src/generated/model/createSponsorshipActivityRequest.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * NUTFes FinanSu API
 * FinanSu APIドキュメント
 * OpenAPI spec version: 2.0.0
 */
import type { ActivityStatus } from './activityStatus';
import type { FeasibilityStatus } from './feasibilityStatus';
import type { DesignProgress } from './designProgress';
import type { CreateSponsorshipActivityRequestSponsorStyleDetailsItem } from './createSponsorshipActivityRequestSponsorStyleDetailsItem';

/**
 * 新規作成時のリクエストボディ
 */
export interface CreateSponsorshipActivityRequest {
  yearPeriodsId: number;
  sponsorId: number;
  userId: number;
  activityStatus: ActivityStatus;
  feasibilityStatus: FeasibilityStatus;
  designProgress: DesignProgress;
  remarks?: string;
  /** 登録したい協賛プラン情報のリスト */
  sponsorStyleDetails?: CreateSponsorshipActivityRequestSponsorStyleDetailsItem[];
}